feat: Rename playwright container package by adding 'internal' prefix
