+++ packages/hypertest-plugin-playwright/package.json
@@ -12,7 +12,7 @@
     "lint": "biome lint"
   },
   "dependencies": {
-    "@hypertest/hypertest-playwright-container": "0.x.x",
+    "@hypertest/internal-hypertest-playwright-container": "0.x.x",
     "@hypertest/hypertest-types": "0.x.x",
     "@playwright/test": "^1.51.1",
     "escape-string-regexp": "^5.0.0",
